+++ api/tests/movies/get-movie.test.ts
@@ -1,39 +1,21 @@
 import { afterAll, beforeAll, describe, expect, it } from "bun:test";
 import { eq } from "drizzle-orm";
-import Elysia from "elysia";
-import { base } from "~/base";
-import { movies } from "~/controllers/movies";
 import { seedMovie } from "~/controllers/seed/movies";
 import { db } from "~/db";
 import { shows } from "~/db/schema";
 import { bubble } from "~/models/examples";
-
-const app = new Elysia().use(base).use(movies);
-const getMovie = async (id: string, langs?: string) => {
-	const resp = await app.handle(
-		new Request(`http://localhost/movies/${id}`, {
-			method: "GET",
-			headers: langs
-				? {
-						"Accept-Language": langs,
-					}
-				: {},
-		}),
-	);
-	const body = await resp.json();
-	return [resp, body] as const;
-};
+import { getMovie } from "./movies-helper";
+import { expectStatus } from "tests/utils";
 
 let bubbleId = "";
 
-function expectStatus(resp: Response, body: object) {
-	const matcher = expect({ ...body, status: resp.status });
-	return {
-		toBe: (status: number) => {
-			matcher.toMatchObject({ status: status });
-		},
-	};
-}
+beforeAll(async () => {
+	const ret = await seedMovie(bubble);
+	bubbleId = ret.id;
+});
+afterAll(async () => {
+	await db.delete(shows).where(eq(shows.slug, bubble.slug));
+});
 
 describe("Get movie", () => {
 	it("Retrive by slug", async () => {
@@ -93,12 +75,4 @@ describe("Get movie", () => {
 		});
 		expect(resp.headers.get("Content-Language")).toBe("en");
 	});
-});
-
-beforeAll(async () => {
-	const ret = await seedMovie(bubble);
-	bubbleId = ret.id;
-});
-afterAll(async () => {
-	await db.delete(shows).where(eq(shows.slug, bubble.slug));
 });
